requirements.txt
airflow==1.7.1.3
alembic==0.8.7
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
attrs==16.3.0
Babel==1.3
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.22
bitarray==0.8.1
breadability==0.1.20
celery==3.1.20
certifi==2016.2.28
cffi==1.9.1
chardet==2.3.0
chartkick==0.4.2
constantly==15.1.0
croniter==0.3.12
cryptography==1.7.2
cssselect==1.0.1
decorator==4.0.9
decorators==0.1.1
dill==0.2.5
dj-database-url==0.4.2
Django==1.11.7
django-extensions==1.6.1
django-heroku==0.2.0
django-rest-framework-mongoengine==3.3.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
djangorestframework-jsonapi==2.0.0b2
djangorestframework-jwt==1.11.0
docker==2.0.2
docker-pycreds==0.2.1
docopt==0.6.2
docutils==0.12
endpoints==1.1.24
enum34==1.1.6
Flask==0.10.1
Flask-Admin==1.4.0
Flask-Cache==0.13.1
Flask-Login==0.2.11
Flask-WTF==0.12
flower==0.8.3
funcsigs==0.4
future==0.15.2
futures==3.0.5
gunicorn==19.3.0
hive-thrift-py==0.0.1
idna==2.2
impyla==0.13.8
incremental==16.10.1
inflection==0.3.1
ipaddress==1.0.18
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
jobtastic==0.3.1
jsonfield==1.0.3
jsonpath-rw==1.4.0
kombu==3.0.33
ldap3==0.9.9
lockfile==0.12.2
lxml==3.5.0
Mako==1.0.4
Markdown==2.6.6
MarkupSafe==0.23
mongoengine==0.10.5
mysqlclient==1.3.6
nltk==3.1
numpy==1.11.1
pandas==0.18.1
parsel==1.1.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
ply==3.8
prompt-toolkit==1.0.7
protorpc==0.12.0
psutil==3.4.2
psycopg2==2.6.1
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
Pygments==2.1.3
PyHive==0.2.1
PyHyphen==2.0.5
PyJWT==1.5.3
pymongo==3.2.1
pyOpenSSL==16.2.0
python-daemon==2.1.1
python-dateutil==2.4.2
python-editor==1.0.1
pytz==2017.3
queuelib==1.4.2
redis==2.10.5
requests==2.9.1
Scrapy==1.3.0
service-identity==16.0.0
setproctitle==1.1.10
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.9
stompy==0.2.9
sumy==0.4.0
textatistic==0.0.1
thrift==0.9.3
tornado==4.3
traitlets==4.3.0
Twisted==16.6.0
unicodecsv==0.14.1
uWSGI==2.0.11.2
virtualenv==13.1.2
w3lib==1.16.0
wcwidth==0.1.7
websocket-client==0.40.0
Werkzeug==0.11.10
whitenoise==3.3.1
WTForms==2.1
zope.interface==4.3.3
